requirements.txt
llama-index==0.10.32
python-dotenv==1.0.1
llama-index-embeddings-huggingface==0.2.0